+++ DataServer/cmd/api-schema-gen/main.go
@@ -54,7 +54,6 @@ func main() {
 		{"SubmitVoiceover", apiwire.SubmitVoiceover{}},
 		{"SubmitSubtitles", apiwire.SubmitSubtitles{}},
 		{"SubmitLayer", apiwire.SubmitLayer{}},
-		{"SubmitSubtitleTrack", apiwire.SubmitSubtitleTrack{}},
 		{"SubmitDeliveryPlanEntry", apiwire.SubmitDeliveryPlanEntry{}},
 		{"SubmitPublishingTarget", apiwire.SubmitPublishingTarget{}},
 		{"SubmitManifestRef", apiwire.SubmitManifestRef{}},